refactor: Redis > Mongo

+++ api/tests/external/test_99_cleanup.py
@@ -19,21 +19,22 @@ def test_singouins_pc_delete():
     response = requests.get(url, headers=headers)
     pcs      = json.loads(response.text)['payload']
 
-    for pc in pcs:
-        if pc['instance'] is not None:
-            # We need to leave the instance
-            url = f"{API_URL}/mypc/{pc['id']}/instance/{pc['instance']}/leave"
-            response = requests.post(url, headers=headers)
+    if pcs and len(pcs) > 0:
+        for pc in pcs:
+            if pc['instance'] is not None:
+                # We need to leave the instance
+                url = f"{API_URL}/mypc/{pc['id']}/instance/{pc['instance']}/leave"
+                response = requests.post(url, headers=headers)
+
+                assert response.status_code == 200
+                assert json.loads(response.text)['success'] is True
+
+            # We delete the PC
+            url        = f"{API_URL}/mypc/{pc['id']}"  # DELETE
+            response   = requests.delete(url, headers=headers)
 
             assert response.status_code == 200
             assert json.loads(response.text)['success'] is True
-
-        # We delete the PC
-        url        = f"{API_URL}/mypc/{pc['id']}"  # DELETE
-        response   = requests.delete(url, headers=headers)
-
-        assert response.status_code == 200
-        assert json.loads(response.text)['success'] is True
 
 
 def test_singouins_auth_delete():
@@ -41,14 +42,6 @@ def test_singouins_auth_delete():
     response = requests.post(url, json=AUTH_PAYLOAD)
     token    = json.loads(response.text)['access_token']
     headers  = {"Authorization": f"Bearer {token}"}
-
-    url      = f'{API_URL}/auth/forgotpassword'  # POST
-    response = requests.post(url,
-                             json={'mail': USER_NAME},
-                             headers=headers)
-
-    assert response.status_code == 200
-    assert 'Password replacement OK' in json.loads(response.text)['msg']
 
     url      = f'{API_URL}/auth/delete'  # DELETE
     response = requests.delete(url,
